Flattening of structure to move folders.
