+++ lib/aggregate-snapshot.js
