+++ lib/aggregate.js
@@ -5,10 +5,10 @@
  *   |___|\_/\___|_||_\__|___/\__,_|\_,_\__\___|   This code is GPL v2.0 licenced.
  **/
 'use strict';
-const AggregateSnapshot = require('../aggregate-snapshot');
-const Event = require('../event');
+const AggregateSnapshot = require('./aggregate-snapshot');
+const Event = require('./event');
 const sprintf = require('sprintf-js').sprintf;
-const validationHelpers = require('../validation-helpers');
+const validationHelpers = require('./validation-helpers');
 
 /**
  * An EventSauce aggregate is the root of the event-sourcing hierarchy. It is the type that
