+++ lib/event.js
@@ -6,7 +6,7 @@
  **/
 'use strict';
 
-const validationHelpers = require('../validation-helpers');
+const validationHelpers = require('./validation-helpers');
 
 /**
  * Describes an event for an aggregate
