+++ lib/memory-repository.js
@@ -6,7 +6,7 @@
  **/
 'use strict';
 
-const Repository = require('../repository');
+const Repository = require('./repository');
 const sprintf = require('sprintf-js').sprintf;
 
 /**
