+++ lib/validation-helpers.js
